Rename (and split) + remove 'imposible' case
